requirements.txt
fastapi
uvicorn
faiss-cpu
openai
python-dotenv
pydantic